UI Canvas 도입

+++ Game/ui/Slider.cpp
@@ -6,18 +6,18 @@ Slider::Slider(std::shared_ptr<I2DRenderable> uiRenderer, TextureIndex image, PO
 	:	mBase(std::make_unique<UIModel>(uiRenderer, image)),
 		mSlider(std::make_unique<UIModel>(uiRenderer, image))
 {
-	mBase->GetUIRect().LTx = LT.x;
-	mBase->GetUIRect().LTy = LT.y;
-	mBase->GetUIRect().width = width;
-	mBase->GetUIRect().height = height;
+	mBase->GetUIRect().LTx = static_cast<float>(LT.x);
+	mBase->GetUIRect().LTy = static_cast<float>(LT.y);
+	mBase->GetUIRect().width = static_cast<float>(width);
+	mBase->GetUIRect().height = static_cast<float>(height);
 
 
 	mUnit = width / 100.f;
 
-	mSlider->GetUIRect().LTx = LT.x;
-	mSlider->GetUIRect().LTy = LT.y;
+	mSlider->GetUIRect().LTx = static_cast<float>(LT.x);
+	mSlider->GetUIRect().LTy = static_cast<float>(LT.y);
 	mSlider->GetUIRect().width = mUnit * mValue;
-	mSlider->GetUIRect().height = height;
+	mSlider->GetUIRect().height = static_cast<float>(height);
 }
 
 Slider::~Slider()
